Add tests for new logic
