+++ UnitTests/Layout+HelpersTests.swift
@@ -1,0 +1,17 @@
+//
+//  Copyright © FINN.no AS, Inc. All rights reserved.
+//
+
+import Foundation
+import XCTest
+@testable import Troika
+
+class LayoutHelpersTests: XCTestCase {
+    func testEdgeInsets() {
+        let inset = EdgeInsets(top: -10, leading: -10, bottom: -10, trailing: -10)
+        XCTAssertEqual(inset.top, 10)
+        XCTAssertEqual(inset.leading, 10)
+        XCTAssertEqual(inset.bottom, 10)
+        XCTAssertEqual(inset.trailing, 10)
+    }
+}
