--- UnitTests/UnitTests.swift
@@ -1,8 +1,0 @@
-//
-//  Copyright © FINN.no AS, Inc. All rights reserved.
-//
-
-import XCTest
-
-class UnitTests: XCTestCase {
-}
